test(profile): move behavioural coverage to component tier
Keep 4 PG happy-path smokes in e2e (view, edit, change-password,
oauth set-password); move login-redirect/validation/set-password
branches to tests/component over a FakeUnitOfWork.
